requirements.txt
matplotlib>=3.10.3
numpy>=2.2.6
opencv-python>=4.11.0.86
pandas>=2.2.3
pillow>=11.2.1
streamlit>=1.45.1
torch>=2.7.0 